Adicionando a lista a IDE RubyMine
Foi atualizado o script shell que valida os diretórios das IDEs

+++ scripts/reset-jetbrains-evaluation-license.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 # Script que faz o reset do evalutaion license do PhpStorm
 JETBRAINS_CONFIG_DIR="$HOME/.config/JetBrains/"
+
 if [ -d "$JETBRAINS_CONFIG_DIR" ]; then
 	echo 'Removendo a licença trial...'
 	sleep 3
@@ -9,9 +10,12 @@ if [ -d "$JETBRAINS_CONFIG_DIR" ]; then
 	echo '# ps - Para remover a licença do PhpStorm                                     #'
 	echo '# ws - Para remover a licença do WebStorm                                     #'
 	echo '# dg - Para remover a licença do DataGrip                                     #'
+	echo '# rm - Para remover a licença do RubyMine                                     #'
 	echo '# sair - Para sair das opções e a removoção padrão do PhpStorm será executada #'
 	echo '##############################################################################'
+
 	LAST_VERSION_CONFIG=$(find "$HOME/.config/JetBrains/" -maxdepth 1 -type d -name "PhpStorm2021.2*" | sort -V | tail -1)
+
 	while :
 	do
         read INPUT_STRING
@@ -28,6 +32,10 @@ if [ -d "$JETBRAINS_CONFIG_DIR" ]; then
                 LAST_VERSION_CONFIG=$(find "$HOME/.config/JetBrains/" -maxdepth 1 -type d -name "DataGrip2021.2*" | sort -V | tail -1)
                 break
                 ;;
+			rm)
+                LAST_VERSION_CONFIG=$(find "$HOME/.config/JetBrains/" -maxdepth 1 -type d -name "RubyMine2021.2*" | sort -V | tail -1)
+                break
+                ;;
             sair)
                 echo "Removendo a licença trail do PhpStorm..."
                 sleep 3
@@ -38,20 +46,30 @@ if [ -d "$JETBRAINS_CONFIG_DIR" ]; then
                 ;;
         esac
     done
+
+	if [ -d "$LAST_VERSION_CONFIG/eval" ]; then
+		echo "#1 Removendo -> $LAST_VERSION_CONFIG/eval"
+		rm -rf "$LAST_VERSION_CONFIG/eval"
+		sleep 2
+		echo "#2 Removendo -> $LAST_VERSION_CONFIG/options/other.xml"
+		rm -rf "$LAST_VERSION_CONFIG/options/other.xml"
+		sleep 2
+	fi
+
 	JETBRAINS_CONSENTOPTIONS=$(find "$HOME/.local/share/JetBrains/" -maxdepth 1 -type d -name "consentOptions*" | sort -V | tail -1)
 	JAVA_USERPREFS=$(find "$HOME/.java/" -maxdepth 1 -type d -name ".userPrefs*" | sort -V | tail -1)
-	echo "#1 Removendo -> $LAST_VERSION_CONFIG/eval"
-	rm -rf "$LAST_VERSION_CONFIG/eval"
-	sleep 2
-	echo "#2 Removendo -> $LAST_VERSION_CONFIG/options/other.xml"
-	rm -rf "$LAST_VERSION_CONFIG/options/other.xml"
-	sleep 2
-	echo "#3 Removendo -> $JETBRAINS_CONSENTOPTIONS"
-	rm -rf "$JETBRAINS_CONSENTOPTIONS"
-	sleep 2
-	echo "#4 Removendo -> $JAVA_USERPREFS"
-	rm -rf "$JAVA_USERPREFS"
-	sleep 2
+
+	if [ -d "$JETBRAINS_CONSENTOPTIONS" ]; then
+		echo "#3 Removendo -> $JETBRAINS_CONSENTOPTIONS"
+		rm -rf "$JETBRAINS_CONSENTOPTIONS"
+		sleep 2
+	fi
+	if [ -d "$JAVA_USERPREFS" ]; then
+		echo "#4 Removendo -> $JAVA_USERPREFS"
+		rm -rf "$JAVA_USERPREFS"
+		sleep 2
+	fi
+
 	echo "."
 	sleep 1
 	echo ".."
